still allow using duckdb without copy to

+++ bmsdna/lakeapi/context/df_duckdb.py
@@ -17,6 +17,7 @@ from datetime import datetime, timezone
 from bmsdna.lakeapi.core.config import SearchConfig
 from uuid import uuid4
 
+ENABLE_COPY_TO = os.environ.get("ENABLE_COPY_TO", "1") == "1"
 
 def _get_temp_table_name():
     return "temp_" + str(uuid4()).replace("-","")
@@ -63,6 +64,8 @@ class DuckDBResultData(ResultData):
         return self.df.fetch_record_batch(chunk_size)
 
     def write_parquet(self, file_name: str):
+        if not ENABLE_COPY_TO:
+            return super().write_parquet(file_name)
         query = get_sql(self.original_sql)
         uuidstr = _get_temp_table_name()
         # temp table required because of https://github.com/duckdb/duckdb/issues/7616
@@ -70,18 +73,24 @@ class DuckDBResultData(ResultData):
         self.con.execute(full_query)
 
     def write_nd_json(self, file_name: str):
+        if not ENABLE_COPY_TO:
+            return super().write_nd_json(file_name)
         query = get_sql(self.original_sql)
         uuidstr = _get_temp_table_name()
         full_query = f"CREATE TEMP VIEW {uuidstr} AS {query}; COPY (SELECT *FROM {uuidstr}) TO '{file_name}' (FORMAT JSON); DROP VIEW {uuidstr}"
         self.con.execute(full_query)
 
-    def write_csv(self, file_name: str, *, separator: str):
+    def write_csv(self, file_name: str, *, separator: str):        
+        if not ENABLE_COPY_TO:
+            return super().write_csv(file_name, separator=separator)
         query = get_sql(self.original_sql)
         uuidstr = _get_temp_table_name()
         full_query = f"CREATE TEMP VIEW {uuidstr} AS {query};  COPY (SELECT *FROM {uuidstr}) TO '{file_name}' (FORMAT CSV, delim '{separator}', header True); DROP VIEW {uuidstr}"
         self.con.execute(full_query)
 
-    def write_json(self, file_name: str):
+    def write_json(self, file_name: str):        
+        if not ENABLE_COPY_TO:
+            return super().write_json(file_name)
         query = get_sql(self.original_sql)
         uuidstr = _get_temp_table_name()
         full_query = f"CREATE TEMP VIEW {uuidstr} AS {query}; COPY (SELECT *FROM {uuidstr})  TO '{file_name}' (FORMAT JSON, Array True); DROP VIEW {uuidstr}"
